Use ATM service to deposit money in Vending Machine.
Add current balance to track how much money was inserted by user.
Reorganize methods in VendingMachine class to alphabetical order
(except for display method left on end).

 Change Product classes to use BigDecimal instead of Double for correct
 money calculation.

 Test a simple buy scenario.

+++ src/main/java/tdd/vendingMachine/products/Beer.java
@@ -1,14 +1,16 @@
 package tdd.vendingMachine.products;
 
+import java.math.BigDecimal;
+
 /**
  * Beer implementation of Vending Machine Product.
  */
 public class Beer implements Product {
 
-    private double price = 4.3;
+    private BigDecimal price = BigDecimal.valueOf(4.3);
 
     @Override
-    public double price() {
+    public BigDecimal price() {
         return price;
     }
