+++ src/main/java/tdd/vendingMachine/products/Cola.java
@@ -1,14 +1,16 @@
 package tdd.vendingMachine.products;
 
+import java.math.BigDecimal;
+
 /**
  * Cola implementation of Vending Machine Product.
  */
 public class Cola implements Product {
 
-    private double price = 2.5;
+    private BigDecimal price = BigDecimal.valueOf(2.5);
 
     @Override
-    public double price() {
+    public BigDecimal price() {
         return price;
     }
