+++ src/main/java/tdd/vendingMachine/products/Product.java
@@ -1,10 +1,12 @@
 package tdd.vendingMachine.products;
 
+import java.math.BigDecimal;
+
 /**
  * Vending Machine product interface for use in factory.
  */
 public interface Product {
 
-    double price();
+    BigDecimal price();
 
 }
