+++ src/main/java/tdd/vendingMachine/products/Water.java
@@ -1,14 +1,16 @@
 package tdd.vendingMachine.products;
 
+import java.math.BigDecimal;
+
 /**
  * Water implementation of Vending Machine Product.
  */
 public class Water implements Product {
 
-    private double price = 1.1;
+    private BigDecimal price = BigDecimal.valueOf(1.1);
 
     @Override
-    public double price() {
+    public BigDecimal price() {
         return price;
     }
